reminder/requirements.txt
requests
supabase
python-dotenv
